Move FusePadAvgPool to canonicalization
