+++ src/Pass/Passes.hpp
@@ -61,7 +61,6 @@ std::unique_ptr<mlir::Pass> createQDQAroundOpOptONNXToONNXPass();
 
 std::unique_ptr<mlir::Pass> createQDQOptONNXToONNXPass();
 std::unique_ptr<mlir::Pass> createFoldDQBinaryQPass();
-std::unique_ptr<mlir::Pass> createFusePadIntoAvgpoolPass();
 
 std::unique_ptr<mlir::Pass> createQuantTypesPass();
