+++ src/Tools/onnx-mlir-opt/RegisterPasses.cpp
@@ -91,10 +91,6 @@ void registerOMPasses(int optLevel) {
       []() -> std::unique_ptr<mlir::Pass> { return createInstrumentPass(); });
 
   mlir::registerPass([]() -> std::unique_ptr<mlir::Pass> {
-    return createFusePadIntoAvgpoolPass();
-  });
-
-  mlir::registerPass([]() -> std::unique_ptr<mlir::Pass> {
     return createInstrumentCleanupPass();
   });
